working changes

+++ CProject/UWFSSD/utility.h
@@ -1,7 +1,7 @@
 #ifndef UTILITY_H
 #define UTILITY_H
 
-void initializeString(char * string, size_t stringLength);
+void initializeString(char * string, int stringLength);
 void generateUUID(char * uuid);
 
 #endif
